Create beforeCAFE.py
This script is use for filtering the Orthogroups.GeneCount.tsv file outputted by Orthofinder. It removed single species gene families and gene families with any one species showing a gene count above 100 to get account for large size differentials. 

Additionally it provides the total number of orthologs and genes in the final output